third_party/blink/web_tests/external/wpt/html/semantics/popups/popup-focus.tentative.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Popup focus behaviors</title>
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popUp.research.explainer">
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/popup-utils.js"></script>

<div popup data-test='default behavior - pop-up is not focused' data-no-focus>
  <p>This is a pop-up</p>
  <button>first button</button>
</div>

<div popup data-test='autofocus pop-up' autofocus tabindex=-1 class=should-be-focused>
  <p>This is a pop-up</p>
</div>

<div popup data-test='autofocus empty pop-up' autofocus tabindex=-1 class=should-be-focused></div>

<div popup data-test='autofocus pop-up with button' autofocus tabindex=-1 class=should-be-focused>
  <p>This is a pop-up</p>
  <button>button</button>
</div>

<div popup data-test='autofocus child'>
  <p>This is a pop-up</p>
  <button autofocus class=should-be-focused>autofocus button</button>
</div>

<div popup data-test='autofocus on tabindex=0 element'>
  <p autofocus tabindex=0 class=should-be-focused>This is a pop-up with autofocus on a tabindex=0 element</p>
  <button>button</button>
</div>

<div popup data-test='autofocus multiple children'>
  <p>This is a pop-up</p>
  <button autofocus class=should-be-focused>autofocus button</button>
  <button autofocus>second autofocus button</button>
</div>

<div popup autofocus tabindex=-1 data-test='autofocus pop-up and multiple autofocus children' class=should-be-focused>
  <p>This is a pop-up</p>
  <button autofocus>autofocus button</button>
  <button autofocus>second autofocus button</button>
</div>

<div popup delegatesfocus data-test='delegatesfocus pop-up'>
  <p>This is a pop-up</p>
  <button class=should-be-focused>first button should be focused</button>
  <button>second button</button>
</div>

<div popup delegatesfocus data-test='delegatesfocus takes precedence over autofocus'>
  <p>This is a pop-up</p>
  <button class=should-be-focused>first button</button>
  <button autofocus>autofocus button should NOT be focused</button>
</div>

<div popup delegatesfocus autofocus data-test='delegatesfocus takes precedence over autofocus 2'>
  <p>This is a pop-up</p>
  <button class=should-be-focused>first button</button>
  <button>autofocus button should NOT be focused</button>
</div>

<div popup delegatesfocus data-test='delegatesfocus on empty pop-up has no effect' data-no-focus></div>

<div popup data-test='delegatesfocus on child has no effect' data-no-focus>
  <p>This is a pop-up</p>
  <button delegatesfocus>first button</button>
</div>

<div popup delegatesfocus data-test='delegatesfocus skips contained pop-ups'>
  <p>This is a pop-up</p>
  <div popup>
    <button>Contained pop-up button</button>
  </div>
  <button class=should-be-focused>first button</button>
  <button>autofocus button should NOT be focused</button>
</div>

<div popup delegatesfocus data-test='delegatesfocus skips contained dialogs'>
  <p>This is a pop-up</p>
  <dialog>
    <button>Contained dialog button</button>
  </dialog>
  <button class=should-be-focused>first button</button>
  <button>autofocus button should NOT be focused</button>
</div>

<style>
  [popup] {
    border: 2px solid black;
    top:150px;
    left:150px;
  }
  :focus-within { border: 5px dashed red; }
  :focus { border: 5px solid lime; }
</style>

<script>
  function addInvoker(t, popUp) {
    const button = document.createElement('button');
    button.innerText = 'Click me';
    const popUpId = 'pop-up-id';
    assert_equals(document.querySelectorAll('#' + popUpId).length, 0);
    document.body.appendChild(button);
    t.add_cleanup(function() {
      popUp.removeAttribute('id');
      button.remove();
    });
    popUp.id = popUpId;
    button.setAttribute('togglepopup', popUpId);
    return button;
  }
  function addPriorFocus(t) {
    const priorFocus = document.createElement('button');
    document.body.appendChild(priorFocus);
    t.add_cleanup(() => priorFocus.remove());
    return priorFocus;
  }
  function activateAndVerify(popUp) {
    const testName = popUp.getAttribute('data-test');
    promise_test(async t => {
      const priorFocus = addPriorFocus(t);
      let expectedFocusedElement = popUp.matches('.should-be-focused') ? popUp : popUp.querySelector('.should-be-focused');
      if (popUp.hasAttribute('data-no-focus')) {
        expectedFocusedElement = priorFocus;
      }
      assert_true(!!expectedFocusedElement);
      assert_false(popUp.matches(':top-layer'));

      // Directly show and hide the pop-up:
      priorFocus.focus();
      assert_equals(document.activeElement, priorFocus);
      popUp.showPopup();
      assert_equals(document.activeElement, expectedFocusedElement, `${testName} activated by popUp.showPopup()`);
      popUp.hidePopup();
      assert_equals(document.activeElement, priorFocus, 'prior element should get focus on hide');

      // Hit Escape:
      priorFocus.focus();
      assert_equals(document.activeElement, priorFocus);
      popUp.showPopup();
      assert_equals(document.activeElement, expectedFocusedElement, `${testName} activated by popUp.showPopup()`);
      await sendEscape();
      assert_equals(document.activeElement, priorFocus, 'prior element should get focus after Escape');

      // Change the pop-up type:
      priorFocus.focus();
      popUp.showPopup();
      assert_equals(document.activeElement, expectedFocusedElement, `${testName} activated by popUp.showPopup()`);
      assert_equals(popUp.popUp, 'auto', 'All pop-ups in this test should start as popup=auto');
      popUp.popUp = 'hint';
      assert_false(popUp.matches(':top-layer'), 'Changing the pop-up type should hide the pop-up');
      assert_equals(document.activeElement, priorFocus, 'prior element should get focus when the type is changed');
      popUp.popUp = 'auto';

      // Remove from the document:
      priorFocus.focus();
      popUp.showPopup();
      assert_equals(document.activeElement, expectedFocusedElement, `${testName} activated by popUp.showPopup()`);
      popUp.remove();
      assert_false(popUp.matches(':top-layer'), 'Removing the pop-up should hide it');
      if (!popUp.hasAttribute('data-no-focus')) {
        assert_not_equals(document.activeElement, priorFocus, 'prior element should *not* get focus when the pop-up is removed from the document');
      }
      document.body.appendChild(popUp);

      // Show a modal dialog:
      priorFocus.focus();
      popUp.showPopup();
      assert_equals(document.activeElement, expectedFocusedElement, `${testName} activated by popUp.showPopup()`);
      const dialog = document.body.appendChild(document.createElement('dialog'));
      dialog.showModal();
      assert_false(popUp.matches(':top-layer'), 'Opening a modal dialog should hide the pop-up');
      assert_not_equals(document.activeElement, priorFocus, 'prior element should *not* get focus when a modal dialog is shown');
      dialog.close();
      dialog.remove();

      // Use an activating element:
      const button = addInvoker(t, popUp);
      priorFocus.focus();
      button.click();
      assert_true(popUp.matches(':top-layer'));
      assert_equals(document.activeElement, expectedFocusedElement, `${testName} activated by button.click()`);

      // Make sure we can directly focus the (already open) pop-up:
      popUp.focus();
      assert_equals(document.activeElement, popUp.hasAttribute('tabindex') ? popUp : expectedFocusedElement, `${testName} directly focus with popUp.focus()`);
      button.click(); // Button is set to toggle the pop-up
      assert_false(popUp.matches(':top-layer'));
      assert_equals(document.activeElement, priorFocus, 'prior element should get focus on button-toggled hide');
    }, "Popup focus test: " + testName);

    promise_test(async t => {
      const priorFocus = addPriorFocus(t);
      assert_false(popUp.matches(':top-layer'), 'pop-up should start out hidden');
      let button = addInvoker(t, popUp);
      assert_equals(button.getAttribute('togglepopup'), popUp.id, 'This test assumes the button uses `togglepopup`.');
      assert_not_equals(button, priorFocus, 'Stranger things have happened');
      assert_false(popUp.contains(button), 'Start with a non-contained button');
      priorFocus.focus();
      assert_equals(document.activeElement, priorFocus);
      popUp.showPopup();
      assert_true(popUp.matches(':top-layer'));
      await clickOn(button); // This will not light dismiss, but will "toggle" the popUp.
      assert_false(popUp.matches(':top-layer'));
      assert_equals(document.activeElement, priorFocus, 'Focus should return to prior focus');

      // Same thing, but the button is contained within the pop-up
      button.removeAttribute('togglepopup');
      button.setAttribute('hidepopup', popUp.id);
      popUp.appendChild(button);
      priorFocus.focus();
      popUp.showPopup();
      assert_true(popUp.matches(':top-layer'));
      if (!popUp.hasAttribute('data-no-focus')) {
        assert_not_equals(document.activeElement, priorFocus, 'focus should shift for this element');
      }
      await clickOn(button);
      assert_false(popUp.matches(':top-layer'), 'clicking button should hide the pop-up');
      assert_equals(document.activeElement, priorFocus, 'Contained button should return focus to the previously focused element');

      // Same thing, but the button is unrelated (no togglepopup)
      button = document.createElement('button');
      document.body.appendChild(button);
      priorFocus.focus();
      popUp.showPopup();
      assert_true(popUp.matches(':top-layer'));
      await clickOn(button); // This will light dismiss the pop-up, focus the prior focus, then focus this button.
      assert_false(popUp.matches(':top-layer'), 'clicking button should hide the pop-up (via light dismiss)');
      assert_equals(document.activeElement, button, 'Focus should go to unrelated button on light dismiss');
    }, "Popup button click focus test: " + testName);

    promise_test(async t => {
      if (popUp.hasAttribute('data-no-focus')) {
        // This test only applies if the pop-up changes focus
        return;
      }
      const priorFocus = addPriorFocus(t);
      assert_false(popUp.matches(':top-layer'), 'pop-up should start out hidden');

      // Move the prior focus out of the document
      priorFocus.focus();
      popUp.showPopup();
      assert_true(popUp.matches(':top-layer'));
      const newFocus = document.activeElement;
      assert_not_equals(newFocus, priorFocus, 'focus should shift for this element');
      priorFocus.remove();
      assert_equals(document.activeElement, newFocus, 'focus should not change when prior focus is removed');
      popUp.hidePopup();
      assert_not_equals(document.activeElement, priorFocus, 'focused element has been removed');
      document.body.appendChild(priorFocus); // Put it back

      // Move the prior focus inside the (already open) pop-up
      priorFocus.focus();
      popUp.showPopup();
      assert_true(popUp.matches(':top-layer'));
      assert_false(popUp.contains(priorFocus), 'Start with a non-contained prior focus');
      popUp.appendChild(priorFocus); // Move inside the pop-up
      assert_true(popUp.contains(priorFocus));
      assert_true(popUp.matches(':top-layer'), 'pop-up should stay open');
      popUp.hidePopup();
      assert_not_equals(document.activeElement, priorFocus, 'focused element is display:none inside the pop-up');
      document.body.appendChild(priorFocus); // Put it back
    }, "Popup corner cases test: " + testName);
  }

  document.querySelectorAll('body > [popup]').forEach(popUp => activateAndVerify(popUp));
</script>
