third_party/blink/web_tests/external/wpt/html/semantics/popups/popup-focus.tentative-expected.txt

This is a testharness.js-based test.
FAIL Popup focus test: default behavior - popup is not focused assert_equals: default behavior - popup is not focused directly focus with popup.focus() expected Element node <div popup="popup" data-test="default behavior - popup is... but got Element node <button id="prior-focus"></button>
FAIL Popup focus test: autofocus popup assert_equals: autofocus popup activated by popup.showPopup() expected Element node <div popup="popup" data-test="autofocus popup" autofocus=... but got Element node <button id="prior-focus"></button>
FAIL Popup focus test: autofocus empty popup assert_equals: autofocus empty popup activated by popup.showPopup() expected Element node <div popup="popup" data-test="autofocus empty popup" auto... but got Element node <button id="prior-focus"></button>
FAIL Popup focus test: autofocus popup with button assert_equals: autofocus popup with button activated by popup.showPopup() expected Element node <div popup="popup" data-test="autofocus popup with button... but got Element node <button id="prior-focus"></button>
FAIL Popup focus test: autofocus child assert_equals: autofocus child activated by popup.showPopup() expected Element node <button autofocus="" class="should-be-focused">autofocus ... but got Element node <button id="prior-focus"></button>
FAIL Popup focus test: autofocus on tabindex=0 element assert_equals: autofocus on tabindex=0 element activated by popup.showPopup() expected Element node <p autofocus="" tabindex="0" class="should-be-focused">Th... but got Element node <button id="prior-focus"></button>
FAIL Popup focus test: autofocus multiple children assert_equals: autofocus multiple children activated by popup.showPopup() expected Element node <button autofocus="" class="should-be-focused">autofocus ... but got Element node <button id="prior-focus"></button>
FAIL Popup focus test: autofocus popup and multiple autofocus children assert_equals: autofocus popup and multiple autofocus children activated by popup.showPopup() expected Element node <div popup="popup" autofocus="" data-test="autofocus popu... but got Element node <button id="prior-focus"></button>
FAIL Popup focus test: delegatesfocus popup assert_equals: delegatesfocus popup activated by popup.showPopup() expected Element node <button class="should-be-focused">first button should be ... but got Element node <button id="prior-focus"></button>
FAIL Popup focus test: delegatesfocus takes precedence over autofocus assert_equals: delegatesfocus takes precedence over autofocus activated by popup.showPopup() expected Element node <button class="should-be-focused">first button</button> but got Element node <button id="prior-focus"></button>
FAIL Popup focus test: delegatesfocus takes precedence over autofocus 2 assert_equals: delegatesfocus takes precedence over autofocus 2 activated by popup.showPopup() expected Element node <button class="should-be-focused">first button</button> but got Element node <button id="prior-focus"></button>
PASS Popup focus test: delegatesfocus on empty popup has no effect
FAIL Popup focus test: delegatesfocus on child has no effect assert_equals: delegatesfocus on child has no effect directly focus with popup.focus() expected Element node <div popup="popup" data-test="delegatesfocus on child has... but got Element node <button id="prior-focus"></button>
FAIL Popup focus test: delegatesfocus skips contained popups assert_equals: delegatesfocus skips contained popups activated by popup.showPopup() expected Element node <button class="should-be-focused">first button</button> but got Element node <button id="prior-focus"></button>
FAIL Popup focus test: delegatesfocus skips contained dialogs assert_equals: delegatesfocus skips contained dialogs activated by popup.showPopup() expected Element node <button class="should-be-focused">first button</button> but got Element node <button id="prior-focus"></button>
Harness: the test ran to completion.
